requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
pillow==12.1.1
sqlparse==0.5.5
tzdata==2025.3
